+++ AiteBar/AppSettingsWindow.xaml.cs
@@ -382,6 +382,9 @@ public partial class AppSettingsWindow : DarkWindow
         ChkShowPresetColorPicker.IsChecked = _settings.ShowPresetColorPicker;
         ChkShowPresetQuickNote.IsChecked = _settings.ShowPresetQuickNote;
         ChkShowPresetQRCodeGenerator.IsChecked = _settings.ShowPresetQRCodeGenerator;
+        ChkShowPresetShowDesktop.IsChecked = _settings.ShowPresetShowDesktop;
+        ChkShowPresetAppsFolder.IsChecked = _settings.ShowPresetAppsFolder;
+        ChkShowPresetCopilot.IsChecked = _settings.ShowPresetCopilot;
         ChkShowTaskbarPositionIndicator.IsChecked = _settings.ShowTaskbarPositionIndicator.GetValueOrDefault(true);
         ChkCheckForUpdatesEnabled.IsChecked = _settings.CheckForUpdatesEnabled;
         _selectedUiCulture = LocalizationService.NormalizeCultureName(_settings.UiCulture);
@@ -560,6 +563,9 @@ public partial class AppSettingsWindow : DarkWindow
         _settings.ShowPresetColorPicker = ChkShowPresetColorPicker.IsChecked ?? false;
         _settings.ShowPresetQuickNote = ChkShowPresetQuickNote.IsChecked ?? false;
         _settings.ShowPresetQRCodeGenerator = ChkShowPresetQRCodeGenerator.IsChecked ?? false;
+        _settings.ShowPresetShowDesktop = ChkShowPresetShowDesktop.IsChecked ?? false;
+        _settings.ShowPresetAppsFolder = ChkShowPresetAppsFolder.IsChecked ?? false;
+        _settings.ShowPresetCopilot = ChkShowPresetCopilot.IsChecked ?? false;
         _settings.ShowTaskbarPositionIndicator = ChkShowTaskbarPositionIndicator.IsChecked ?? true;
         _settings.CheckForUpdatesEnabled = ChkCheckForUpdatesEnabled.IsChecked ?? true;
         _settings.UiCulture = _selectedUiCulture;